Added button to start and stop execution
Also changed the name of the "end_startup" variable to "running" as it is just a boolian denoting if the game of life is currently running. Additionally fixed a bug where if the borders on the right or left were clicked or dragged over an out of bounds array access error would occur.